requirements.txt
Babel==2.5.3
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.5.0
Django==1.11.12
django-appconf==1.0.2
django-debug-toolbar==1.9.1
django-extensions==2.0.7
django-extra-views==0.6.4
django-geoip==0.5.2.1
django-haystack==2.8.1
django-oscar==1.5.3
django-oscar-api==1.3.1
django-phonenumber-field==1.3.0
django-progressive-web-app==0.1.1
django-select2==6.0.2
django-shortuuidfield==0.1.3
django-tables2==1.16.0
django-treebeard==4.3
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
environ==1.0
factory-boy==2.10.0
Faker==0.8.13
gunicorn==19.8.1
idna==2.6
mock==2.0.0
pbr==4.0.2
phonenumberslite==8.9.4
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
purl==1.4
pycountry==17.9.23
pysolr==3.7.0
python-dateutil==2.7.2
pytz==2018.4
requests==2.18.4
shortuuid==0.5.0
six==1.11.0
sorl-thumbnail==12.4.1
sqlparse==0.2.4
text-unidecode==1.2
Unidecode==0.4.21
urllib3==1.22
whitenoise==3.3.1
